Run bazel installer as a argument to bash
We used to set the installer as executable on X20 and have kokoro copy it to
the temp folder and run it from there. Now that executables on x20 must be
built verifiable, we cannot set the +x bit any more.

Instead, run the script as an argument to the bash command.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=204536872