Check gradle wrapper in to VCS
The version of gradle used in compilation is now controlled by the gradle
wrapper and is centrally managed. This means that everyone will be using the
same version of gradle for build, including IC systems like Travis and Cloud
Build.

The Gradle version for this Gradle wrapper is v5.0.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=223856122