Reduce # of manual instances in production and sandbox
Steady state QPS for prod is <30 on both default and pubapi, so number of instances can easily be brought safely down to 15 & 10 instances.  Sandbox has negligible steady state QPS (especially on pubapi), so it is brought even lower.  Note that, if we have any issues with these levels, we can increase them instantly using the nomulus set_num_instances command, without having to do another release or a rollback; this is therefore low risk.

Note that we'll want to go back to 100 instances for the first day of .dev sunrise as well as the entire week of EAP and first day of GA.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=226566822