Move domain registry terraform configs
We are moving toward using GitHub as the source of truth for the domain registry project (Nomulus). As such, the piper location will soon be deleted, along with it the terraform configs. These files are copied to the canonical location []

Note that the files under modules will still be present in the open source code base as it allows open source users to set up the project quickly. The files under envs are specific to each actual project and is removed entirely from the open source code (it was excluded by MOE before).

Some files are renamed to conform to the newly established terraform code style.

There was a remaining regarding using latchkey to set up IAM policies that I intend to punt for now. I imagine if we decide to use latchkey, it means that the IAM related terraform configs will be removed for the Annealing set up. However we would still like to leave that in the open source configs such that it still is a one-stop shop to set up your project.

The automation mode is set to DRYRUN so that there are no accidental changes to our projects during .dev launch. It will be changed back later.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=234838043