LSC: Mark Python binaries with explicit python_version.
Having PY2 global version default is deprecated and change in 2019Q3.  This LSC ([] is a prerequisite to that Blaze change to prevent unintended breakages when we flip the default.

This CL is no-op to “freeze” the world. If it conflicts with your plans to move to PY3, feel free to change forward to PY3 as you see fit (eg. remove python_version when you move to py2and3_test).

BEGIN_PUBLIC
n/a
END_PUBLIC

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=245028632